refactor: simplify subpool
