+++ crates/transaction-pool/src/pool/parked.rs
@@ -1,10 +1,12 @@
 use crate::{identifier::TransactionId, TransactionOrdering, ValidPoolTransaction};
 use fnv::FnvHashMap;
-use reth_primitives::U256;
 use std::{cmp::Ordering, collections::BTreeSet, sync::Arc};
 
-/// A Sub-Pool that currently holds transactions that violate the dynamic fee requirement
-pub(crate) struct BaseFeePool<T: TransactionOrdering> {
+/// A pool of transaction that are currently parked and wait for external changes that eventually
+/// move the transaction into the pending pool.
+///
+/// This pool is a bijection: at all times each set contains the same transactions.
+pub(crate) struct ParkedPool<T: TransactionOrdering> {
     /// How to order transactions.
     ordering: Arc<T>,
     /// Keeps track of transactions inserted in the pool.
@@ -12,14 +14,14 @@ pub(crate) struct BaseFeePool<T: TransactionOrdering> {
     /// This way we can determine when transactions where submitted to the pool.
     submission_id: u64,
     /// _All_ Transactions that are currently inside the pool grouped by their identifier.
-    by_id: FnvHashMap<TransactionId, Arc<BasFeeTransaction<T>>>,
+    by_id: FnvHashMap<TransactionId, Arc<QueuedPoolTransaction<T>>>,
     /// All transactions sorted by their priority function.
-    best: BTreeSet<Arc<BasFeeTransaction<T>>>,
+    best: BTreeSet<Arc<QueuedPoolTransaction<T>>>,
 }
 
-// === impl BaseFeePool ===
+// === impl QueuedPool ===
 
-impl<T: TransactionOrdering> BaseFeePool<T> {
+impl<T: TransactionOrdering> ParkedPool<T> {
     /// Create a new pool instance.
     pub(crate) fn new(ordering: Arc<T>) -> Self {
         Self { ordering, submission_id: 0, by_id: Default::default(), best: Default::default() }
@@ -30,11 +32,8 @@ impl<T: TransactionOrdering> BaseFeePool<T> {
     /// # Panics
     ///
     /// if the transaction is already included
-    pub(crate) fn add_transaction(
-        &mut self,
-        _tx: Arc<ValidPoolTransaction<T::Transaction>>,
-        _basefee: U256,
-    ) {
+    pub(crate) fn add_transaction(&mut self, _tx: Arc<ValidPoolTransaction<T::Transaction>>) {
+        todo!()
     }
 
     /// Removes the transaction from the pool
@@ -54,33 +53,31 @@ impl<T: TransactionOrdering> BaseFeePool<T> {
     }
 }
 
-/// The transaction type of this pool
-struct BasFeeTransaction<T: TransactionOrdering> {
+/// A reference to a transaction in the pool
+struct QueuedPoolTransaction<T: TransactionOrdering> {
     /// Identifier that tags when transaction was submitted in the pool.
     submission_id: u64,
     /// Actual transaction.
     transaction: Arc<ValidPoolTransaction<T::Transaction>>,
     /// The priority value assigned by the used `Ordering` function.
     priority: T::Priority,
-    /// The base fee
-    base_fee: T::Priority,
 }
 
-impl<T: TransactionOrdering> Eq for BasFeeTransaction<T> {}
+impl<T: TransactionOrdering> Eq for QueuedPoolTransaction<T> {}
 
-impl<T: TransactionOrdering> PartialEq<Self> for BasFeeTransaction<T> {
+impl<T: TransactionOrdering> PartialEq<Self> for QueuedPoolTransaction<T> {
     fn eq(&self, other: &Self) -> bool {
         self.cmp(other) == Ordering::Equal
     }
 }
 
-impl<T: TransactionOrdering> PartialOrd<Self> for BasFeeTransaction<T> {
+impl<T: TransactionOrdering> PartialOrd<Self> for QueuedPoolTransaction<T> {
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         Some(self.cmp(other))
     }
 }
 
-impl<T: TransactionOrdering> Ord for BasFeeTransaction<T> {
+impl<T: TransactionOrdering> Ord for QueuedPoolTransaction<T> {
     fn cmp(&self, other: &Self) -> Ordering {
         // This compares by `priority` and only if two tx have the exact same priority this compares
         // the unique `transaction_id`.
